+++ crates/objc2/src/encode/__unstable.rs
@@ -18,7 +18,9 @@
 )]
 
 use crate::encode::{Encode, Encoding};
+use crate::rc::{Id, Ownership};
 use crate::runtime::Bool;
+use crate::Message;
 
 mod return_private {
     pub trait Sealed {}
@@ -28,18 +30,18 @@ mod return_private {
 ///
 /// We currently don't need a similar `EncodeArgument` trait, but we might in
 /// the future.
-pub unsafe trait EncodeReturn: return_private::Sealed {
+pub trait EncodeReturn: return_private::Sealed {
     /// The Objective-C type-encoding for this type.
     const ENCODING_RETURN: Encoding;
 }
 
 impl return_private::Sealed for () {}
-unsafe impl EncodeReturn for () {
+impl EncodeReturn for () {
     const ENCODING_RETURN: Encoding = Encoding::Void;
 }
 
 impl<T: Encode> return_private::Sealed for T {}
-unsafe impl<T: Encode> EncodeReturn for T {
+impl<T: Encode> EncodeReturn for T {
     const ENCODING_RETURN: Encoding = T::ENCODING;
 }
 
@@ -50,49 +52,37 @@ mod convert_private {
 impl<T: EncodeReturn> convert_private::Sealed for T {}
 impl convert_private::Sealed for bool {}
 
+// Implemented in rc/writeback.rs
+impl<T: Message, O: Ownership> convert_private::Sealed for &mut Id<T, O> {}
+impl<T: Message, O: Ownership> convert_private::Sealed for Option<&mut Id<T, O>> {}
+impl<T: Message, O: Ownership> convert_private::Sealed for &mut Option<Id<T, O>> {}
+impl<T: Message, O: Ownership> convert_private::Sealed for Option<&mut Option<Id<T, O>>> {}
+
 /// Represents types that can easily be converted to/from an [`Encode`] type.
 ///
 /// This is implemented specially for [`bool`] to allow using that as
 /// Objective-C `BOOL`, where it would otherwise not be allowed (since they
 /// are not ABI compatible).
+///
+/// This is also done specially for `&mut Id<_, _>`-like arguments, to allow
+/// using those as "out" parameters.
 pub trait EncodeConvertArgument: convert_private::Sealed {
     /// The inner type that this can be converted to and from.
     #[doc(hidden)]
     type __Inner: Encode;
 
+    /// A helper type for out parameters.
     #[doc(hidden)]
-    fn __from_inner(inner: Self::__Inner) -> Self;
+    type __StoredBeforeMessage: Sized;
 
     #[doc(hidden)]
-    fn __into_inner(self) -> Self::__Inner;
-}
+    fn __from_declared_param(inner: Self::__Inner) -> Self;
 
-impl<T: Encode> EncodeConvertArgument for T {
-    type __Inner = Self;
+    #[doc(hidden)]
+    fn __into_argument(self) -> (Self::__Inner, Self::__StoredBeforeMessage);
 
-    #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner
-    }
-
-    #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        self
-    }
-}
-
-impl EncodeConvertArgument for bool {
-    type __Inner = Bool;
-
-    #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner.as_bool()
-    }
-
-    #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        Bool::new(self)
-    }
+    #[doc(hidden)]
+    unsafe fn __process_after_message_send(_stored: Self::__StoredBeforeMessage) {}
 }
 
 /// Same as [`EncodeConvertArgument`], but for return types.
@@ -102,23 +92,55 @@ pub trait EncodeConvertReturn: convert_private::Sealed {
     type __Inner: EncodeReturn;
 
     #[doc(hidden)]
-    fn __from_inner(inner: Self::__Inner) -> Self;
+    fn __into_declared_return(self) -> Self::__Inner;
 
     #[doc(hidden)]
-    fn __into_inner(self) -> Self::__Inner;
+    fn __from_return(inner: Self::__Inner) -> Self;
+}
+
+impl<T: Encode> EncodeConvertArgument for T {
+    type __Inner = Self;
+
+    type __StoredBeforeMessage = ();
+
+    #[inline]
+    fn __from_declared_param(inner: Self::__Inner) -> Self {
+        inner
+    }
+
+    #[inline]
+    fn __into_argument(self) -> (Self::__Inner, Self::__StoredBeforeMessage) {
+        (self, ())
+    }
 }
 
 impl<T: EncodeReturn> EncodeConvertReturn for T {
     type __Inner = Self;
 
     #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner
+    fn __into_declared_return(self) -> Self::__Inner {
+        self
     }
 
     #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        self
+    fn __from_return(inner: Self::__Inner) -> Self {
+        inner
+    }
+}
+
+impl EncodeConvertArgument for bool {
+    type __Inner = Bool;
+
+    type __StoredBeforeMessage = ();
+
+    #[inline]
+    fn __from_declared_param(inner: Self::__Inner) -> Self {
+        inner.as_bool()
+    }
+
+    #[inline]
+    fn __into_argument(self) -> (Self::__Inner, Self::__StoredBeforeMessage) {
+        (Bool::new(self), ())
     }
 }
 
@@ -126,13 +148,13 @@ impl EncodeConvertReturn for bool {
     type __Inner = Bool;
 
     #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner.as_bool()
+    fn __into_declared_return(self) -> Self::__Inner {
+        Bool::new(self)
     }
 
     #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        Bool::new(self)
+    fn __from_return(inner: Self::__Inner) -> Self {
+        inner.as_bool()
     }
 }
 
@@ -149,7 +171,7 @@ mod args_private {
 ///
 /// Note that tuples themselves don't implement [`Encode`] directly, because
 /// they're not FFI-safe!
-pub unsafe trait EncodeArguments: args_private::Sealed {
+pub trait EncodeArguments: args_private::Sealed {
     /// The encodings for the arguments.
     const ENCODINGS: &'static [Encoding];
 }
@@ -158,7 +180,7 @@ macro_rules! encode_args_impl {
     ($($Arg: ident),*) => {
         impl<$($Arg: EncodeConvertArgument),*> args_private::Sealed for ($($Arg,)*) {}
 
-        unsafe impl<$($Arg: EncodeConvertArgument),*> EncodeArguments for ($($Arg,)*) {
+        impl<$($Arg: EncodeConvertArgument),*> EncodeArguments for ($($Arg,)*) {
             const ENCODINGS: &'static [Encoding] = &[
                 // T::__Inner::ENCODING => T::ENCODING
                 // bool::__Inner::ENCODING => Bool::ENCODING
@@ -204,8 +226,11 @@ mod tests {
             TypeId::of::<<i32 as EncodeConvertArgument>::__Inner>(),
             TypeId::of::<i32>()
         );
-        assert_eq!(<i32 as EncodeConvertArgument>::__from_inner(42), 42);
-        assert_eq!(EncodeConvertArgument::__into_inner(42i32), 42);
+        assert_eq!(
+            <i32 as EncodeConvertArgument>::__from_declared_param(42),
+            42
+        );
+        assert_eq!(EncodeConvertArgument::__into_argument(42i32).0, 42);
     }
 
     #[test]
@@ -214,21 +239,25 @@ mod tests {
             TypeId::of::<<i8 as EncodeConvertArgument>::__Inner>(),
             TypeId::of::<i8>()
         );
-        assert_eq!(<i8 as EncodeConvertArgument>::__from_inner(-3), -3);
-        assert_eq!(EncodeConvertArgument::__into_inner(-3i32), -3);
+        assert_eq!(<i8 as EncodeConvertArgument>::__from_declared_param(-3), -3);
+        assert_eq!(EncodeConvertArgument::__into_argument(-3i32).0, -3);
     }
 
     #[test]
     fn convert_bool() {
-        assert!(!<bool as EncodeConvertArgument>::__from_inner(Bool::NO));
-        assert!(<bool as EncodeConvertArgument>::__from_inner(Bool::YES));
-        assert!(!<bool as EncodeConvertReturn>::__from_inner(Bool::NO));
-        assert!(<bool as EncodeConvertReturn>::__from_inner(Bool::YES));
+        assert!(!<bool as EncodeConvertArgument>::__from_declared_param(
+            Bool::NO
+        ));
+        assert!(<bool as EncodeConvertArgument>::__from_declared_param(
+            Bool::YES
+        ));
+        assert!(!<bool as EncodeConvertReturn>::__from_return(Bool::NO));
+        assert!(<bool as EncodeConvertReturn>::__from_return(Bool::YES));
 
-        assert!(!EncodeConvertArgument::__into_inner(false).as_bool());
-        assert!(EncodeConvertArgument::__into_inner(true).as_bool());
-        assert!(!EncodeConvertReturn::__into_inner(false).as_bool());
-        assert!(EncodeConvertReturn::__into_inner(true).as_bool());
+        assert!(!EncodeConvertArgument::__into_argument(false).0.as_bool());
+        assert!(EncodeConvertArgument::__into_argument(true).0.as_bool());
+        assert!(!EncodeConvertReturn::__into_declared_return(false).as_bool());
+        assert!(EncodeConvertReturn::__into_declared_return(true).as_bool());
 
         #[cfg(all(feature = "apple", target_os = "macos", target_arch = "x86_64"))]
         assert_eq!(